+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- CKEditor now requires Craft CMS 5.6+. 
+- CKEditor now requires Craft CMS 5.6+.
 - Added support for viewing Stripe settings when `allowAdminChanges` is disabled. ([#78](https://github.com/craftcms/stripe/pull/78))
 
 ## 1.3.3 - 2025-01-15